[ksqlDb.RestApi.Client]: PropertyInfoExtractor renamed to MemberInfoExtractor
